requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.21
celery==3.1.19
Django==1.8.6
django-bootstrap3==6.2.2
django-celery==3.1.17
django-ckeditor==5.0.2
django-cors-headers==1.2.2
django-registration==2.0.3
django-rest-swagger==2.0.6

# dynamic-rest must come before djangorestframework.
# The SchemaGenerator from DRF used by django-rest-swagger has a bug in 3.4.0
# which is fixed in >=3.4.1. The schema page breaks if this order is reversed.
dynamic-rest==1.5.0
djangorestframework==3.4.6

djangorestframework-jwt==1.8.0
dnspython==1.14.0
google-api-python-client==1.5.2
kombu==3.0.33
lxml==3.5.0
oauth2client==3.0.0
psycopg2==2.6.1
pycrypto==2.6.1
python-dateutil==2.6.0
pytz==2015.7
requests==2.17.3
# Selenium is used for ShoalScrape, not testing.
selenium==2.53.6
uWSGI==2.0.11.2
wheel==0.24.0
